Move pact merger to pacts module, add PACT_PARSING_OPTIONS, do not overwrite different-header requests
